+++ src/components/card-list.tsx
@@ -1,13 +1,48 @@
+'use client';
+
+import { useLiveQuery } from 'dexie-react-hooks';
+import { db } from '@/db';
+
 import { cn } from '@/lib/utils';
-import { qrCodeData } from '@/constants';
+import { qrCodeSeedData } from '@/constants';
+
+import { useEffect, useState } from 'react';
 import { QRCard } from './qr-card';
 import ExtraCards from './extra-cards.temp';
+import { QRCardLoading, QRCardSeed, QRCardSkeleton } from './skeleton-qr-card';
 
 interface cardListProps extends React.HTMLAttributes<HTMLDivElement> {
   className?: string;
 }
 
 const CardList: React.FC<cardListProps> = ({ className }) => {
+  const [loading, setLoading] = useState(true);
+
+  const qrCodeDataFriends = useLiveQuery(async () => {
+    try {
+      const qrs = await db.qrs.toArray();
+
+      return qrs;
+    } catch (error) {
+      console.error(`🚀 Unable to fetch db.qrs in card-list`, error);
+      return [];
+    }
+  }, []);
+
+  useEffect(() => {
+    if (qrCodeDataFriends) {
+      setLoading(false);
+    }
+  }, [qrCodeDataFriends]);
+
+  const seed = async () => {
+    try {
+      await db.qrs.bulkAdd(qrCodeSeedData);
+    } catch (error) {
+      console.error(`🚀 Seeding failed, db.qrs.bulkAdd:`, error);
+    }
+  };
+
   return (
     <div
       className={cn(
@@ -15,7 +50,24 @@ const CardList: React.FC<cardListProps> = ({ className }) => {
         className
       )}
     >
-      {qrCodeData.map((qrCode) => (
+      {/* loading card with text */}
+      {loading && <QRCardLoading />}
+
+      {/* FIXME: skeleton colors not appearing */}
+      {loading && <QRCardSkeleton />}
+
+      {/* No data, seed me */}
+      {!loading && !qrCodeDataFriends?.length && (
+        <QRCardSeed
+          seed={seed}
+          disabled={loading}
+          qrCodeDataLength={qrCodeSeedData.length}
+        />
+      )}
+      {/* Extra skeleton */}
+      <QRCardSkeleton />
+
+      {qrCodeDataFriends?.map((qrCode) => (
         <QRCard
           key={qrCode.type + qrCode.title || qrCode.description || ``}
           {...qrCode}