+++ tailwind.config.js
@@ -1,6 +1,5 @@
 const tailwindCssAnimate = require(`tailwindcss-animate`);
 const { fontFamily } = require(`tailwindcss/defaultTheme`);
-const tailwindLineClamp = require(`@tailwindcss/line-clamp`);
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
@@ -80,5 +79,5 @@ module.exports = {
       },
     },
   },
-  plugins: [tailwindCssAnimate, tailwindLineClamp],
+  plugins: [tailwindCssAnimate],
 };